pig_23320/recursos_api/serializers.py
from administracion.models import Estudiante
from rest_framework import serializers


class EstudianteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Estudiante
        fields = '__all__'
