pig_23320/requirements.txt
asgiref==3.6.0
astroid==2.15.1
colorama==0.4.6
dill==0.3.6
Django==3.2
django-extensions==3.2.1
django-shell==1.0.2
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.6
pycodestyle==2.10.0
pylint==2.17.1
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
typing-extensions==4.5.0
wrapt==1.15.0